// src/api/match/content-types/match/match.ts
// Interface automatically generated by schemas-to-ts

import { Player } from '../../../player/content-types/player/player';
import { Player_Plain } from '../../../player/content-types/player/player';
import { AdminPanelRelationPropertyModification } from '../../../../common/schemas-to-ts/AdminPanelRelationPropertyModification';

export interface Match {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    matchId: string;
    map?: string;
    mapName?: string;
    mode?: string;
    players?: { data: Player[] };
  };
}
export interface Match_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  matchId: string;
  map?: string;
  mapName?: string;
  mode?: string;
  players?: Player_Plain[];
}

export interface Match_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  matchId: string;
  map?: string;
  mapName?: string;
  mode?: string;
  players?: number[];
}

export interface Match_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  matchId: string;
  map?: string;
  mapName?: string;
  mode?: string;
  players?: AdminPanelRelationPropertyModification<Player_Plain>;
}